==> kap1_manuscript_book/18-distance-estimation.Rmd <==
# Distance estimattion

In this analysis, we want to see if negative or positive correlation between KRAB-ZNFs and TEs are related to their distance on chromosome. We only can compare KRAB-ZNFs and TEs on the same chromosome.
1. Extract positive/negative correlation pairs from human data
2. Filter them from coordinates reference
3. caculate distance and save results as a dataframe

```{r, warning=FALSE}
library(twice)
library(dplyr)
library(biomaRt)
```


Here we load the human only correlation results [path need to be updated when revising the book]

```{r}
hmCorr <- read.csv("../hm_results/hmCorrResult81.csv")
hmCorr.sig <- hmCorr %>%
    filter(padj < 0.05)
hmCorr.sig$geneName[1]
hmCorr.sig$teName[1]
```
We then filter them from coordinates reference from biomart reference. You can find tutorial [here](http://grch37.ensembl.org/info/data/biomart/biomart_r_package.html)
```{r}
# get significant znfs and TEs name
znfs <- unique(hmCorr.sig$geneName)
tes <- unique(hmCorr.sig$teName)

# get hg19 coordinates information
grch37 <- useEnsembl(biomart = "ensembl", GRCh=37) 
listDatasets(grch37)

ensembl = useEnsembl(biomart="ensembl", dataset="hsapiens_gene_ensembl", GRCh=37)

ensembl_gene <- getBM(
    attributes = c('ensembl_gene_id',
                   'hgnc_symbol',
                   'chromosome_name',
                   'start_position',
                   'end_position'),
    mart = ensembl
)


# filter znfs
ensembl_gene_znfs <- ensembl_gene[ensembl_gene$hgnc_symbol %in% znfs, ]
ensembl_gene_znfs[42, "chromosome_name"] <- "19" #modified previous unknown result

```

We based on the chromosome number and te name from correlation result to filter some data
```{r}
data("hg19rmsk_gtf")

chr_list <- unique(ensembl_gene_znfs$chromosome_name)

#filter chromosome
df_te <- hg19rmsk_gtf
df_te$seqnames <- substr(df_te$seqnames, 4, length(df_te$seqnames)) # remove chr
df_te <- df_te %>% 
    filter(seqnames %in% chr_list) %>%
    filter(gene_id %in% tes)

```

The inputs are ready, and then we can write a algorithm to calculate their distance based on the same chromosome.
```{r include=FALSE}
# 1. iterate through chr number
# 2. calculate mean value and decide which will be first and second
# 4. calculate distance and add rows to dataframe

# initialize an empty dataframe
df <- data.frame(
    Chr = character(),
    KRAB_ZNFs = character(),
    TEs = character(),
    distance = numeric(),
    pos_or_neg = character()
)

for (chr in chr_list){
    
    dfGene <- ensembl_gene_znfs[ensembl_gene_znfs$chromosome_name==chr, ]
    dfTE <- df_te[df_te$seqnames==chr, ]
    df_idx <- 0
    
    for (i in 1:nrow(dfGene)){
        for (j in 1:nrow(dfTE)){

            geneMean <- mean(c(dfGene$start_position[i], dfGene$end_position[i]))
            teMean <- mean(c(dfTE$start[j], dfTE$end[j]))
            dis <- 0
            
            # check if the data exist in the correlation table
            check_status <- hmCorr.sig[
                hmCorr.sig$geneName==dfGene$hgnc_symbol[i] &
                    hmCorr.sig$teName==dfTE$gene_id[j],
            ]
            
            if (nrow(check_status)>0) {

                if (geneMean > teMean){
                    dis <- dfGene$start_position[i] - dfTE$end[j]
                } else {
                    dis <- dfTE$start[j] - dfGene$end_position[i]
                }

                coef <- hmCorr.sig[hmCorr.sig$geneName==dfGene$hgnc_symbol[i] &
                                       hmCorr.sig$teName==dfTE$gene_id[j], "coef"]
                
                coef_result <- ifelse(coef>0, "pos", "neg")
                
                df_idx <- df_idx + 1
                df[df_idx, ] <- c(
                    chr,
                    dfGene$hgnc_symbol[i],
                    dfTE$gene_id[j],
                    dis,
                    ifelse(coef>0, "pos", "neg")
                )
            }
        }
    }

}

df$distance <- as.numeric(df$distance)

write.table(df, "results/hm/kznfsTEs_distance.csv", sep=",")
```

#### Visualize the results

```{r}
library(ggplot2)
library(dplyr)
library(hrbrthemes)
library(viridis)
```
We can look at each ZNF genes separately. Here we use ZNF782 as example.

```{r}
# ZNF782
znf782 <- df %>%
    filter(KRAB_ZNFs=="ZNF782") %>%
    mutate(distance = log2(distance))

p.znf782 <- znf782 %>%
  ggplot( aes(x=distance, fill=pos_or_neg)) +
    geom_histogram( color="#e9ecef", alpha=0.6, position = 'identity') +
    scale_fill_manual(values=c("#69b3a2", "#404080")) +
    theme_ipsum() +
    labs(fill="")

p.znf782
```



```{r}
p1 <- ggplot(df, aes(x=KRAB_ZNFs, y=distance/1000000, fill=pos_or_neg)) +
    geom_boxplot(alpha=0.5) +
    scale_fill_manual(values=c("#69b3a2", "#f5587f")) +
    theme_ipsum_tw(base_size=8) +
    xlab("KRAB-ZNFs") +
    ylab("distance(bp/1e+06)")
```

Density plot
```{r}
ggplot(df, aes(x=distance/1000000, group=pos_or_neg, fill=pos_or_neg)) +
    geom_density(adjust=1.5, alpha=.4) +
    scale_fill_manual(values=c("#69b3a2", "#f5587f")) +
    theme_ipsum(
        base_size = 16
    ) + 
    xlab("distance(bp/1e+06)")
```

The next step, we try to visualize them using histogram and also combined with data that are not on the same chromosome.

```{r}
df %>%
    ggplot(aes(x=distance/1000000, fill=pos_or_neg)) +
    geom_histogram( color="#e9ecef", alpha=0.4, position = 'identity') +
    scale_fill_manual(values=c("#69b3a2", "#f5587f")) +
    theme_light()
```


```{r}
df %>%
    ggplot(aes(x=distance/1000000, fill=pos_or_neg)) +
    geom_histogram( color="#e9ecef", alpha=0.4, position = 'identity', aes(y=log2(..count.. + 1))) +
    scale_fill_manual(values=c("#69b3a2", "#f5587f")) +
    theme_light()
```

```{r}
p2 <- df %>%
    ggplot(aes(x=distance/1000000, fill=pos_or_neg)) +
    geom_histogram( color="#e9ecef", alpha=0.4, position = 'identity', aes(y=log2(..count.. + 1))) +
    geom_density(aes(y=log2(..count.. + 1)), alpha=.3) +
    scale_fill_manual(values=c("#69b3a2", "#f5587f")) +
    theme_ipsum_tw(base_size=10)
```

Next, we try to investigate without considerate the distance. (1) We pick KZNFs-TEs pairs from the correlation tables (2) Calculate the number from TE reference (3) Remove 4038 data from our result. 
Then we can get how many positive and negative result we have remain.
```{r}
# 1. select data
hmCorr.sig$pos_or_neg <- ifelse(hmCorr.sig$coef>0, "pos", "neg")
df_all <- hmCorr.sig[,c(1,2,6)]
rownames(df_all) <- 1:nrow(df_all)

# 2. calculate number from TE reference

for (i in 1:nrow(df_all)){
    te_name <- df_all$teName[i]
    count <- nrow(hg19rmsk_gtf[hg19rmsk_gtf$gene_id==te_name, ])
    df_all$count[i] <- count
    
}

num_pos <- sum(df_all[df_all$pos_or_neg=="pos", "count"]) #94955
num_neg <- sum(df_all[df_all$pos_or_neg=="neg", "count"]) #100167

# 3.
# from the distance data
num_pos_dis <- nrow(df[df$pos_or_neg=="pos", ]) # 118
num_neg_dis <- nrow(df[df$pos_or_neg=="neg", ]) # 3920

df_remain <- data.frame(
    type = factor(c("pos_same_chr", "neg_same_chr", "pos", "neg"), levels = c("pos_same_chr", "neg_same_chr", "pos", "neg")),
    count = c(num_pos_dis, num_neg_dis, num_pos - num_pos_dis, num_neg - num_neg_dis),
    pos_or_neg = c("pos", "neg", "pos", "neg")
)

p3 <- ggplot(data = df_remain, aes(x=type, y=log2(count+1), fill=pos_or_neg) )+
    geom_bar(stat="identity", alpha=0.6, color="#e9ecef") + 
    scale_fill_manual(values=c("#69b3a2", "#f5587f")) +
    theme_ipsum_tw(base_size=10) +
    theme(axis.text.x = element_text(angle = 45, hjust=1))
    

```

```{r, warning=FALSE}
library(ggpubr)
figure <- ggarrange(
    p1,p2,p3, 
    ncol=3, 
    nrow=1, 
    legend = "bottom",
    labels = c("A", "B", "C")
)

ggsave(figure, height=4, width=12, filename = "figures/distance.jpg", dpi=400, bg="white")
```
This part, we called 'artifact' at the moment because we create rows with artifact KRAB-ZNFs and TEs for infinity distance. We add 94837 rows of positive correlation data and 96247 negative correlation data.

```{r warning=FALSE}
# try to merge
artifact_kznf_pos <- data.frame(
    Chr=rep("X", 94837),
    KRAB_ZNFs = rep("a_KZNF", 94837),
    TEs = rep("a_TE", 94837),
    distance = rep(228845445, 94837),
    pos_or_neg = rep("pos", 94837)
)

artifact_kznf_neg <- data.frame(
    Chr=rep("X", 96247),
    KRAB_ZNFs = rep("b_KZNF", 96247),
    TEs = rep("b_TE", 96247),
    distance = rep(248845445, 96247),
    pos_or_neg = rep("neg", 96247)
)
    

df_artifact <- rbind(df, artifact_kznf_neg, artifact_kznf_pos)

p4 <- df_artifact %>%
    ggplot(aes(x=distance/1000000, fill=pos_or_neg)) +
    geom_histogram( color="#e9ecef", alpha=0.4, position = 'identity', aes(y=log2(..count.. + 1))) +
    scale_fill_manual(values=c("#69b3a2", "#f5587f")) +
    theme_ipsum_tw(base_size=10)

p5 <- df_artifact %>%
    ggplot(aes(x=distance/1000000, fill=pos_or_neg)) +
    geom_histogram( color="#e9ecef", alpha=0.4, position = 'identity', aes(y=log2(..count.. + 1))) +
    geom_density(aes(y=log2(..count.. + 1)), alpha=.3) +
    scale_fill_manual(values=c("#69b3a2", "#f5587f")) +
    theme_ipsum_tw(base_size=10)


figure_artifact <- ggarrange(
    p4, p5, 
    ncol=2, 
    nrow=1, 
    legend = "bottom",
    labels = c("A", "B")
)

ggsave(figure_artifact, height=4, width=8, filename = "figures/distance_option.jpg", dpi=400, bg="white")

```
